MOD int return value changed from double to int

+++ include/libft_addons.h
@@ -6,7 +6,7 @@
 /*   By: jkauppi <jkauppi@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/06 14:53:13 by jkauppi           #+#    #+#             */
-/*   Updated: 2021/03/10 00:41:35 by jkauppi          ###   ########.fr       */
+/*   Updated: 2021/03/10 01:22:28 by jkauppi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ double			ft_radian(double angle_degree);
 int				ft_max_int(int nbr1, int nbr2);
 int				ft_min_int(int nbr1, int nbr2);
 double 			ft_mod_double(double dividend, double divisor);
-double 			ft_mod_int(int dividend, int divisor);
+int 			ft_mod_int(int dividend, int divisor);
 void			ft_matrix_x_vector_double(t_matrix_size matrix_size,
 						double **matrix, double *vector, double *new_vector);
